rename ThemeProperties to asset-registries/Themes
